third_party/WebKit/LayoutTests/platform/android/svg/zoom/page/zoom-get-screen-ctm-expected.txt

This test checks getScreenCTM() on zoomed pages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
FAIL CTM1 should be 1, 0, 0, 1, 0, 100. Was 1, 0, 0, 1, 0, 145.13888549804688.
FAIL CTM2 should be 1, 0, 0, 1, 100, 200. Was 1, 0, 0, 1, 100, 245.13888549804688.
FAIL CTM3 should be 1, 0, 0, 1, 200, 300. Was 1, 0, 0, 1, 200, 345.1388854980469.
FAIL CTM4 should be 1, 0, 0, 1, 300, 400. Was 1, 0, 0, 1, 300, 445.1388854980469.

PASS successfullyParsed is true

TEST COMPLETE
